cobalt/posix: fcntl: turn the generic argument into a long value
In order to prevent unexpected truncation of pointer args in userland
with the LP64 data model, libcobalt's fcntl() wrapper should accept a
long (3rd) argument.

Anticipate this change in the corresponding syscall implementation in
the Cobalt core.  The updated ABI remains backward-compatible for
current users.

+++ kernel/cobalt/posix/io.c
@@ -53,7 +53,7 @@ COBALT_SYSCALL(close, lostage, (int fd))
 	return rtdm_fd_close(fd, 0);
 }
 
-COBALT_SYSCALL(fcntl, current, (int fd, int cmd, int arg))
+COBALT_SYSCALL(fcntl, current, (int fd, int cmd, long arg))
 {
 	return rtdm_fd_fcntl(fd, cmd, arg);
 }
